In that case the login program doesn't create a file to be sourced.
it outputs directly the version.

git-svn-id: http://euclid.esac.esa.int/svn/EC/SGS/SDC/CH/Projects/EuclidEnv/trunk@7126 463500fe-d6be-4184-9b20-47717e75ecac

+++ scripts/ELogin.sh
@@ -5,7 +5,7 @@ python_loc=`python -c "from distutils.sysconfig import get_python_lib; print(get
 
 if [[ -r ${python_loc}/Euclid/Login.py ]]; then
   ELogin_tmpfile=`python ${python_loc}/Euclid/Login.py --shell=sh --mktemp "$@"`
-  ELoginStatus="$?"  
+  ELoginStatus="$?"
 else
   ELogin_tmpfile=`python -m Euclid.Login --shell=sh --mktemp "$@"`
   ELoginStatus="$?"
@@ -16,11 +16,15 @@ unset python_loc
 needs_cleanup=no
 
 if [[ "$ELoginStatus" = 0 && -n "$ELogin_tmpfile" ]]; then
-  . $ELogin_tmpfile
+  if [[ -r $ELogin_tmpfile ]]; then
+    . $ELogin_tmpfile
+    rm -f $ELogin_tmpfile
+  else
+    echo "$ELogin_tmpfile"
+  fi
   needs_cleanup=yes
 fi
 
-rm -f $ELogin_tmpfile
 unset ELogin_tmpfile
 
 if [[ "$needs_cleanup" = "yes" ]]; then
